CR - Move variantMapping into constants
